requirements.txt
flask
gunicorn
openai-whisper
ffmpeg-python
pytest
